Use external package for generating code with razor.
