+++ Qowaiv.sln
@@ -1,11 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0
+VisualStudioVersion = 15.0.27004.2009
 MinimumVisualStudioVersion = 15.0
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Qowaiv", "src\Qowaiv\Qowaiv.csproj", "{ACF6F726-8CBE-4F7C-89EF-9002A6441D19}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Qowaiv.CodeGenerator", "src\Qowaiv.CodeGenerator\Qowaiv.CodeGenerator.csproj", "{E3DCEC1A-818F-4EF1-AECA-2E34E96B39FC}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Qowaiv.UnitTests", "test\Qowaiv.UnitTests\Qowaiv.UnitTests.csproj", "{D0233D1C-86A8-4229-907D-438537AF4372}"
 EndProject
@@ -69,6 +67,12 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Qowaiv.ComponentModel.UnitT
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Qowaiv.ComponentModel", "src\Qowaiv.ComponentModel\Qowaiv.ComponentModel.csproj", "{98F628EF-952A-4487-95D8-C717E855F42C}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tooling", "Tooling", "{CCCCEB32-7082-431D-A7DC-3C6E213DA8D4}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Qowaiv.CodeGenerator", "tooling\Qowaiv.CodeGenerator\Qowaiv.CodeGenerator.csproj", "{DC848917-E057-461F-A8B0-276F3C61CDF8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Qowaiv.CodeGenerator.UnitTests", "test\Qowaiv.CodeGenerator.UnitTests\Qowaiv.CodeGenerator.UnitTests.csproj", "{9A884773-78CD-4753-9E08-22350B2370FA}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -79,10 +83,6 @@ Global
 		{ACF6F726-8CBE-4F7C-89EF-9002A6441D19}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{ACF6F726-8CBE-4F7C-89EF-9002A6441D19}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{ACF6F726-8CBE-4F7C-89EF-9002A6441D19}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E3DCEC1A-818F-4EF1-AECA-2E34E96B39FC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E3DCEC1A-818F-4EF1-AECA-2E34E96B39FC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E3DCEC1A-818F-4EF1-AECA-2E34E96B39FC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E3DCEC1A-818F-4EF1-AECA-2E34E96B39FC}.Release|Any CPU.Build.0 = Release|Any CPU
 		{D0233D1C-86A8-4229-907D-438537AF4372}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D0233D1C-86A8-4229-907D-438537AF4372}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D0233D1C-86A8-4229-907D-438537AF4372}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -111,6 +111,14 @@ Global
 		{98F628EF-952A-4487-95D8-C717E855F42C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{98F628EF-952A-4487-95D8-C717E855F42C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{98F628EF-952A-4487-95D8-C717E855F42C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DC848917-E057-461F-A8B0-276F3C61CDF8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DC848917-E057-461F-A8B0-276F3C61CDF8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DC848917-E057-461F-A8B0-276F3C61CDF8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DC848917-E057-461F-A8B0-276F3C61CDF8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9A884773-78CD-4753-9E08-22350B2370FA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9A884773-78CD-4753-9E08-22350B2370FA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9A884773-78CD-4753-9E08-22350B2370FA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9A884773-78CD-4753-9E08-22350B2370FA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -121,6 +129,8 @@ Global
 		{41C15FF1-5053-4D59-8314-B6E1CE89CDC1} = {A4C3C9FF-4EED-41FE-8849-C896429254B5}
 		{C4A408F0-FCFD-4FB9-AC08-BEA5E25DC2B8} = {A4C3C9FF-4EED-41FE-8849-C896429254B5}
 		{69CA9631-8F91-4213-880D-CBB2D516CC58} = {40C1E942-8DB1-4E58-8FB4-71F5EDA66029}
+		{DC848917-E057-461F-A8B0-276F3C61CDF8} = {CCCCEB32-7082-431D-A7DC-3C6E213DA8D4}
+		{9A884773-78CD-4753-9E08-22350B2370FA} = {40C1E942-8DB1-4E58-8FB4-71F5EDA66029}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {9B55A336-E2D3-4656-9840-D519B3E1159B}
