--- src/Qowaiv.CodeGenerator/CodeDom/Razor/RazorTemplate_T.cs
@@ -1,9 +1,0 @@
-﻿namespace Qowaiv.CodeGenerator.CodeDom.Razor
-{
-    /// <summary>Represents a typed razor template.</summary>
-    public abstract class RazorTemplate<T> : RazorTemplate
-    {
-        /// <summary>The underlying typed Model.</summary>
-        public T Model { get; set; }
-    }
-}
