--- src/Qowaiv.CodeGenerator/Properties/AssemblyInfo.cs
@@ -1,6 +1,0 @@
-﻿using System.Reflection;
-
-// Info about the assembly.
-[assembly: AssemblyTitle("Qowaiv.CodeGenerator")]
-[assembly: AssemblyProduct("Qowaiv.CodeGenerator")]
-[assembly: AssemblyDescription("Commandline tool to generate Qowaiv code")]
