--- src/Qowaiv.CodeGenerator/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="ExcelLibrary" version="1.2011.7.31" targetFramework="net45" />
-  <package id="log4net" version="2.0.5" targetFramework="net45" />
-  <package id="Microsoft.AspNet.Razor" version="3.2.3" targetFramework="net45" />
-</packages>
