+++ test/Qowaiv.CodeGenerator.UnitTests/Qowaiv.CodeGenerator.UnitTests.csproj
@@ -1,0 +1,18 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net45</TargetFramework>
+    <SignAssembly>false</SignAssembly>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="NUnit" Version="3.9.0" />
+    <PackageReference Include="NUnit3TestAdapter" Version="3.9.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\tooling\Qowaiv.CodeGenerator\Qowaiv.CodeGenerator.csproj" />
+    <ProjectReference Include="..\Qowaiv.UnitTests\Qowaiv.UnitTests.csproj" />
+  </ItemGroup>
+
+</Project>
