+++ test/Qowaiv.CodeGenerator.UnitTests/Xml/XResourceFileDataTest.cs
