+++ test/Qowaiv.CodeGenerator.UnitTests/Xml/XResourceFileHeaderTest.cs
