+++ test/Qowaiv.CodeGenerator.UnitTests/Xml/XResourceFileTest.cs
