+++ tooling/Qowaiv.CodeGenerator/Generators/Svo/Templates/SvoStruct_JavaScript.cshtml
@@ -1,0 +1,87 @@
+﻿/**
+ * Represents @Model.a @(Model.ClassLongName).
+ * 
+ * @@constructor
+ * @@remarks Should not be used directly, use @(Model.ClassName).create().
+ */
+@(Model.ClassName) = function (){ this.v = null };
+
+@(Model.ClassName).prototype = {
+    /**
+     * The underlying value.
+     */
+    v: @Model.JavaScriptInitialValue,
+    /**
+     * Compares the object with an other object.
+     * @@param {object} other The object to compare with.
+     * @@return True if the other object equals this object, otherwise false.
+     */
+    eq: function (other) {
+        return other instanceof @(Model.ClassName) && this.v === other.v;
+    },
+    /** 
+     * Returns true if the @(Model.ClassLongName) represents an empty value, otherwise false.
+     */
+    isEmpty: function(){
+        return this.v === @Model.JavaScriptInitialValue;
+    },
+    /** 
+     * Returns a string that represents the current @(Model.ClassLongName).
+     */
+    toString: function () {
+        return v === null ? '' : v.toString();
+    },
+    /** 
+     * Returns a value representing the current @(Model.ClassLongName) for JSON.
+     * @@remarks Is used by JSON.stringify().
+     */
+    toJSON: function () {
+        return this.v;
+    },
+    /** 
+     * Returns a formatted string that represents the current @(Model.ClassLongName).
+     * @@param {string} f The format that this describes the formatting.
+     * @@param {string} c The culture to format to.
+     */
+    format: function (f, c) {
+        return this.toString();
+    }
+};
+
+/** 
+ * The pattern describing a @(Model.ClassLongName).
+ */
+@(Model.ClassName).pattern = /ComplexPattern/;
+
+/** 
+ * Returns true if the specified value is a valid @(Model.ClassLongName), otherwise false.
+ * @@param {object} v A value representing @(Model.ClassLongName).
+ * @@param {string} c The culture to falidate for.
+ */
+@(Model.ClassName).isValid = function (v, c) {
+    return @(Model.ClassName).create(v, c) !== null;
+};
+
+/**
+ * Creates @Model.a @(Model.ClassLongName).
+ * @@param {object} v A value representing @(Model.ClassLongName).
+ * @@param {string} c The culture to create for.
+ * @@return @(Model.aUpper) @(Model.ClassLongName) if valid, otherwise null.
+ */
+@(Model.ClassName).create = function (v, c) {
+    if (v !== null && v !== undefined) {
+        var res = new @(Model.ClassName)();
+        res.v = v;
+        return res;
+    }
+    return null;
+};
+
+/**
+ * Creates @(Model.a) @(Model.ClassLongName) from a JSON string.
+ * @@param {object} v A value representing @(Model.ClassLongName).
+ * @@return @(Model.aUpper) @(Model.ClassLongName) if valid, otherwise null.
+ */
+@(Model.ClassName).fromJSON = function (v) {
+    return @(Model.ClassName).create(v);
+};
