+++ tooling/Qowaiv.CodeGenerator/Guard.cs
