+++ tooling/Qowaiv.CodeGenerator/Qowaiv.CodeGenerator.csproj
@@ -1,0 +1,56 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <Import Project="..\..\build\common.props" />
+  
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net45</TargetFramework>
+    <SignAssembly>false</SignAssembly>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <EmbeddedResource Include="..\..\LICENSE.txt" Link="LICENSE.txt" />
+    <EmbeddedResource Include="Generators\Svo\Templates\SvoStruct.cshtml" />
+    <EmbeddedResource Include="Generators\Svo\Templates\SvoStruct_JavaScript.cshtml" />
+    <EmbeddedResource Include="Generators\Svo\Templates\SvoStruct_TypeConverter.cshtml" />
+    <EmbeddedResource Include="Generators\Svo\Templates\SvoStruct_UnitTest.cshtml" />
+    <EmbeddedResource Include="Resources\Country.xls" />
+    <EmbeddedResource Include="Resources\CountryToCurrency.xls" />
+    <EmbeddedResource Include="Resources\Currency.xls" />
+    <EmbeddedResource Include="Resources\Gender.xls" />
+    <EmbeddedResource Include="Resources\IBAN.xls" />
+    <EmbeddedResource Include="Resources\PostalCode.xls" />
+    <EmbeddedResource Include="Resources\Unknown.xls" />
+    <EmbeddedResource Include="Resources\Web\InternetMediaType.xls" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="ExcelLibrary" Version="1.2011.7.31" />
+    <PackageReference Include="log4net" Version="2.0.8" />
+    <PackageReference Include="RazorEngine" Version="3.10.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\Qowaiv\Qowaiv.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Reference Include="System.Windows.Forms" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Compile Update="QowaivCodeGeneratorMessages.Designer.cs">
+      <DesignTime>True</DesignTime>
+      <AutoGen>True</AutoGen>
+      <DependentUpon>QowaivCodeGeneratorMessages.resx</DependentUpon>
+    </Compile>
+  </ItemGroup>
+
+  <ItemGroup>
+    <EmbeddedResource Update="QowaivCodeGeneratorMessages.resx">
+      <Generator>ResXFileCodeGenerator</Generator>
+      <LastGenOutput>QowaivCodeGeneratorMessages.Designer.cs</LastGenOutput>
+    </EmbeddedResource>
+  </ItemGroup>
+
+</Project>
