+++ tooling/Qowaiv.CodeGenerator/Resources/Currency.xls
