+++ tooling/Qowaiv.CodeGenerator/Xml/XResourceFileHeader.cs
